requirements.txt
django
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
certifi==2020.6.20
defusedxml==0.6.0
diff-match-patch==20200713
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1596543024352/work
django-import-export==2.3.0
django-pandas==0.6.2
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1596215850360/work
odfpy==1.4.1
openpyxl==3.0.4
pandas @ file:///C:/ci/pandas_1596824377390/work
psycopg2 @ file:///C:/ci/psycopg2_1594305172475/work
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
